Move GA4 related components to the analytics module.
