+++ assets/js/modules/analytics-4/components/common/index.js
@@ -16,6 +16,5 @@
  * limitations under the License.
  */
 
-export { default as ActivateSwitch } from './ActivateSwitch';
 export { default as PropertySelect } from './PropertySelect';
 export { default as UseSnippetSwitch } from './UseSnippetSwitch';
