+++ assets/js/modules/analytics/components/common/index.js
@@ -32,5 +32,6 @@ export { default as PropertySelect } from './PropertySelect';
 export { default as PropertySelectIncludingGA4 } from './PropertySelectIncludingGA4';
 export { default as TrackingExclusionSwitches } from './TrackingExclusionSwitches';
 export { default as UseSnippetSwitch } from './UseSnippetSwitch';
+export { default as GA4ActivateSwitch } from './GA4ActivateSwitch';
 export { default as GA4PropertyNotice } from './GA4PropertyNotice';
 export { default as ZeroDataMessage } from './ZeroDataMessage';
