+++ assets/js/modules/analytics/components/settings/index.js
@@ -20,3 +20,4 @@ export { default as SettingsEdit } from './SettingsEdit';
 export { default as SettingsForm } from './SettingsForm';
 export { default as SettingsView } from './SettingsView';
 export { default as SettingsControls } from './SettingsControls';
+export { default as GA4SettingsControls } from './GA4SettingsControls';
